Split up FactoryMapperTest
